Title: 체질량지수 2, Time: 146ms, Memory: 12MB, Status: Wrong Answer - CodeTree

+++ 240830/체질량지수 2/bmi-2.kt
@@ -2,6 +2,6 @@ fun main()
 {
     val (cm,kg)=readLine()!!.split(" ").map{it.toInt()}
     val m=cm.toDouble()/100
-    println(String.format("%.0f",kg/(m*m)))
+    println(String.format("%.0f",kg.toDouble()/(m*m)))
     if(kg/(m*m)>=25)print("Obesity")
 }
